Redesigned reading and displaying of events. Previous design did not have a lossless in memory representation of events, and could not elegantly display dependent events. Implemented an in memory representation and implemented contructors for it. Contructors are completely untested and will not yet compile because of other redesigning issues.
Factored LogFile into LogDoc (the document controller) and LogData (the log file model).

Because of change in representation of events, LogView is no longer compatible with LogFile (which does not exist)
or with LogData (which is new).
Need to change LogView to read data from LogData.  Implementation should be more staighforward than last time because
the data is now better represented.



git-svn-id: https://smlnj-gforge.cs.uchicago.edu/svn/manticore/trunk@4015 ac4fbcad-d21d-0410-b760-c7b1d3380187