adding a specialized ZeroLimitPtr function for preemption by the RTS
the reason for this is that I noticed VProcPreempt has the following sequence
when using SetLimitPtr(vp, 0)

; rdi = vp
xorl   %eax, %eax
xchgq  %rax, 0x200(%rdi)

and I felt that it really should be xorq  %rax, %rax; since I don't think
it's safe to assume the higher-order bits are already clear.

In the end, this didn't fix the bug I wanted to, and I have no evidence
that this fixes any bug at all, but it's safer to ensure that it's 0ULL
since we would otherwise not be preempted and instead overrun our heap!