fix a conceptual mistake in the Major GC scanning routine.
During a Major collection, pointers are updated in two ways:

  1. if they're in the older part of the Major heap, the object
     is forwarded to the Global heap's to-space
  2. if they're in the youngest part of the Major heap, we simply
     adjust the pointer by a fixed amount, because we will later
     memmove that whole area data so it remains in the Major heap.

Thus, when the Major GC scans the global to-space to continue
updating pointers within those objects, we must use the majorGC
scanning routines from the table. The "ScanGlobalToSpace" functions
are really routines that are used to promote all pointers to the
global heap, which the Major GC specifically does not do.