remove optimization to avoid tenuring young data
Previously, after a minor gc we would track data in the major heap
based on how young it is, i.e., whether it was the most recently
allocated data in the major heap.

Any data that was just moved to the major heap in the prior minor
collection, would be left behind during a major collection to avoid
tenuring that known live & young data to the global heap.

This optimization adds complexity to the GC implementation for a
an expected non-negative, but unknown performance benefit relative
to the maintenance cost.

In addition, the merge2.pml regression test fails on all
stack-based compiles due to some issue with stale pointers that
was traced back to this optimization, and is now fixed by this
removal.

It should be possible to reimplement this optimization correctly
in the context of stack-based compiles. It is just not a priority
right now.