Reimplement wrap-captures for arbitrary throws to local conts.
The expression we wrap may have arbirary throws to
local conts within the function it previously belonged to.

Previously we just hoped this didn't happen and if it did,
would raise an error. Now we generate a proper landing pad
that can dispatch to more conts than the return or escape cont.