Seg stacks should not free global seg. on underflow.
It's only safe to free a global segment during the
synchronized global GC.

The cost of freeing the segment if we were to use
a thread-safe data structure (say, a global pool
of allocated segments) would likely negate the
benefit of doing this optimization, since underflow
should be fast.

One potential optimization would be to check if the
segment is managed by (in the allocated list of) the
current vproc. In that case it would be fine to free.