fixed a huge but subtle GC bug in the LLVM backend
It turns out that the LHS type of an alloc may not match up with
the types of the values on the RHS in terms of the types with respect
to whether the type is a GC pointer or not.

For example, if we have:

val l : [any] = alloc(nil : enum(0))

then the GC header tag would treat it as a RAW, because we expected
that the nil value would be of type any to be equal the LHS, i.e.,
no implicit casting here.
Turns out we have to get the header tag from the type of the result
of the alloc, as this patch does now.