[linkstack] merge heap checks into the prologue
This commit adds basic functionality for merging
(or "absorbing") heap limit tests into the prologue,
since we already often need to do one there too.

Currently, the choice of which heap checks to absorb
into the prologue uses a simple heuristic: look for
a heap limit test in the entry CFG block.

In the future we can get more fancy with
absorb-heap-tests's heuristic to pick up more
cases, such as

    S
   / \
  A   B

Where S is the start block, and

(1) least one of the blocks A or B has a heap test.
(2) the blocks with a heap test are not part of a cycle.

Since we know the start block has no incoming edge from the
CFG within the function, part (2) doesn't apply to it and
so we can quickly check for and optimize that case.