added a significant optimization for segmented & resize stacks
Bruggeman et al's segment cache-management adds the current segment
to the cache in the event of both an underflow-return and
during a continuation-throw, since the throw is always to a different
segment.

This patch adds freeing in the latter event (throws), which has a very nice
effect on running times (fib(30) with callec from 2.25 sec -> 0.08 sec).