fix problem with joins not being promoted to gotos
If the binding of a Goto or Other continuation occurs
before the use of a Join continuation, then that
use is considered to happen in a different function environment.

This fact was only tracked by ClassifyConts for uses within the body
of the cont, but not in the expression following the binding.
This must happen for the latter case too because of a situation such
as:

cont joinK () = ... in
  cont escapingK () = ..E.. in
    throw joinK()

Previously such a violation would only be detected if the throw
was in E.