enable nosealingcapture optimization on throw for contig
Contiguous stacks are always non-sealing captures, so the
same optimization used for seg and resizing stacks on throw
to eagarly free the from segment applies, and nets a 1.44x
speed-up on cml-pingpong, making it now faster than seg
and resizing stacks.

This was realized while preparing the submission for pldi but
did not have time to try before deadline.
will need to update the figures in a camera-ready version, etc.