solved perf issue with hybrid stacks
in ec-sudan it was very slow because the segment
size was basically stuck at the minimum given
how often callec occurs and how recursive it is.

so, now whenever we need to grow the stack
and a callec has occured, preventing copying,
we jump to the normal size of a segment
and act like a segmented stack.