fix performance of resizing stacks on ec-* tests
The old impl just used the default stack size when
overflowing from a non-copyable segment. This is
actually a poor decision and instead when using
nocopyoverflow, we double the size (without copying).

This wasn't noticed in previous versions of the RTS probably
due to bugs that were fixed in this branch related to promotions.