third shot at fixing the CPS contract
This time I'm doing it right. I tried to just get away with
having the expansive inliner handle it for us since it
does the analysis already. Problem is that it relies on a
function-size heuristic to stop iterating. So it can blow
up the program and cause a compiler hang after it finishes
when trying to compute CFA on such a big program for seq-boyer.