fix overclassification of conts as GotoCont
Suppose you have:

cont localK () = ... in
  if ...
     then cont escapeK() =
            ...
            throw localK ()
          in
             store escapeK
             ...
      else
          throw localK

Previous analysis and wrapping would consider localK
to be a GotoCont because it has a throw who's context
is within an non Join/Return cont. Thus, the translation
of this code would involve _two_ callec's. One for
localK and one for escapeK. This has knock-on affects
and causes other conts to be classified similarly
and introduce needless callec's.

The newer version of wrap-captures can handle this case
more efficiently so that we don't need so many callecs,
and it will dispatch to the right continuation instead.

Thus, the bug is that the analysis was not updated
accordingly once wrap-captures was upgraded a while back.

This fixes problems with a space leak in cml-spawn where
a tail-call for the iteration loop turns into a non-tail call,
which caused stack overflows for contig stacks in relation to
the number of loop iterations performed.