over 5x speedup in cml-spawn for contig stacks
profiling suggested that ReclaimStacks was the slowest
part of this benchmark, with loading the next pointer of
of the StackInfo linked-list being the most sampled point.

A long time ago I thought it would be better to put the StackInfo
in the stack area's allocation itself, adjacent to the base of the
stack.

This appears to be a very bad decision for cache locality, it
turns out! I guess that part of memory is cold and when you have
tons of stacks, those StackInfo objects in memory are forcably
spread out all over the place!

This optimization simply uses a separate malloc call to allocate
the StackInfo, allowing malloc to place these objects on the
same page of memory. This alone improved the perf of allocation-heavy
cml-spawn from 2.4 seconds down to 0.457 for 128MB contig stacks!